DynamicPricingAPI/requirements.txt
Flask
pandas
tensorflow-cpu
joblib